requirements.txt
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
awscli==1.18.28
awsebcli==3.17.1
bcrypt==3.1.7
beautifulsoup4==4.9.0
blessed==1.17.2
boto3==1.12.17
botocore==1.15.31
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1
colorama==0.3.9
cryptography==2.8
cssselect==1.1.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
dominate==2.5.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.16.0
google-search==1.0.2
idna==2.7
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lexical-diversity==0.1.1
lexicalrichness==0.1.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.4
pandas==1.0.3
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.1.2
psycopg2==2.8.5
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pylint==2.5.2
PyNaCl==1.3.0
pynytimes==0.2.1
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.2
regex==2020.4.4
requests==2.20.1
requests-file==1.5.1
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
semantic-version==2.5.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
spacy>=2.2.0,<3.0.0
SQLAlchemy==1.3.9
termcolor==1.1.0
textblob==0.15.3
texttable==1.6.2
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.0
tqdm==4.46.0
typed-ast==1.4.1
urllib3==1.24.3
vaderSentiment==3.3.1
visitor==0.1.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
wget==3.2
wikipedia==1.4.0
Wikipedia-API==0.5.4
wrapt==1.12.1
WTForms==2.3.1
yamllint==1.20.0
yanytapi==2.0.0
zipp==3.1.0
/opt/python/ondeck/app/en_core_web_sm-2.2.5.tar.gz

